+++ packages/upload-api/src/blob.js
@@ -1,6 +1,7 @@
 import { blobAddProvider } from './blob/add.js'
 import { blobListProvider } from './blob/list.js'
 import { blobRemoveProvider } from './blob/remove.js'
+import { blobReplicateProvider } from './blob/replicate.js'
 import { blobGetProvider } from './blob/get.js'
 import * as API from './types.js'
 
@@ -20,6 +21,7 @@ export function createService(context) {
     add: blobAddProvider(context),
     list: blobListProvider(context),
     remove: blobRemoveProvider(context),
+    replicate: blobReplicateProvider(context),
     get: {
       0: {
         1: blobGetProvider(context),