+++ packages/upload-api/src/test/handlers/plan.spec.js
@@ -1,3 +1,0 @@
-import * as Plan from './plan.js'
-import { test } from '../test.js'
-test({ 'plan/*': Plan.test })