+++ packages/w3up-client/src/account.js
@@ -80,6 +80,7 @@ export const list = ({ agent }, { account } = {}) => {
  * @param {EmailAddress} email
  * @param {object} [options]
  * @param {AbortSignal} [options.signal]
+ * @param {API.AppName} [options.appName]
  * @returns {Promise<API.Result<Account, Error>>}
  */
 export const login = async ({ agent }, email, options = {}) => {
@@ -105,6 +106,7 @@ export const login = async ({ agent }, email, options = {}) => {
     {
       account,
       access: Access.accountAccess,
+      appName: options.appName,
     }
   )