+++ packages/w3up-client/src/types.ts
@@ -20,6 +20,7 @@ import { type Client } from './client.js'
 import { StorefrontService } from '@storacha/filecoin-client/storefront'
 export * from '@ucanto/interface'
 export * from '@storacha/did-mailto'
+export { AppName } from '@storacha/access/types'
 export type { Agent, CapabilityQuery } from '@storacha/access/agent'
 export type {
   Access,