+++ packages/ui/packages/react/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.7.0 (2025-05-21)
+
+### 🚀 Features
+
+- Add support for directing users to bsky.storage after plan selection. ([#267](https://github.com/storacha/upload-service/pull/267))
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/ui-core to 2.4.52
+
+### ❤️ Thank You
+
+- Travis Vachon
+
 ## 2.6.11 (2025-05-13)
 
 ### 🧱 Updated Dependencies